mod filesystem fat32

+++ src/ext4.c
@@ -14,7 +14,7 @@
 
 char * navigateFileExtentTree(int fd, uint64_t offset, uint16_t blockSize, struct ext_super_block sb, uint64_t tableOffset, uint64_t fileSize, uint64_t * curSize);
 
-char * getFileFragment(int fd, uint64_t offset, uint16_t size);
+char * getFileFragment(int fd, uint64_t offset, uint64_t size);
 
 uint32_t navigateDirExtentTree(int fd, char *filename, uint64_t offset, uint16_t blockSize, struct ext_super_block sb, uint64_t tableOffset);
 
@@ -213,7 +213,7 @@ char * navigateFileExtentTree(int fd, uint64_t offset, uint16_t blockSize, struc
 
 }
 
-char * getFileFragment(int fd, uint64_t offset, uint16_t size) {
+char * getFileFragment(int fd, uint64_t offset, uint64_t size) {
 
     //Check size
     if(size == 0) {
